Change Reaction Context to be able to use cached message or download new
Testing Module for the test commands
